server/requirements.txt
fastapi
openai
pydantic-settings
sqlmodel
uvicorn[standard]